feat: extend deterministic symbol attribution to CLI secrets findings
Closes the remaining gap: find_symbol_at_location was Flash-Review-only,
but the "file, line, symbol" claim also applies to `aletheore scan`'s
secrets findings. Moved the function to aletheore.evidence_resolution
(the canonical home for every other location->evidence resolver, e.g.
_resolve_symbol, resolve_recent_commit) so both the CLI and github-app
share one implementation instead of two - flash_review.py now imports
it rather than defining its own copy.

scan_repository (evidence.py) attaches finding["symbol"] to every
working-tree secrets finding right after find_secrets runs, using the
module graph that's already built by that point. Deliberately NOT
applied to history_findings: those carry no line number at all (a
diff hunk, not a location), and even if they did, attaching the
CURRENT tree's symbol name to a bygone commit's line would describe
code that may have since moved, been renamed, or been deleted.

dead_code.py and vulnerabilities.py findings are out of scope, not
gaps: dead-code findings are whole-file ("no other module imports
this file"), and vulnerability findings are keyed on
ecosystem+package+advisory_id (a dependency-manifest fact) - neither
has a line-level location to attach a symbol to.

Verified additive: air_schema.py's _SECRET_FINDING check only requires
path/line/pattern, and its own docstring says findings arrays
deliberately allow extra detector-added keys.
